Apply some other tweaks
