+++ crates/ibc/src/mock/ics18_relayer/mod.rs
@@ -2,4 +2,3 @@
 
 pub mod context;
 pub mod error;
-pub mod utils;
